pyproject.toml
[tool.poetry]
name = "django_server"
version = "1.0"
description = ""
authors = ["taeju.kim <iam@taeju.kim>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
gunicorn = "^20.1.0"
mysqlclient = "^2.1.1"
pysqlite3 = "^0.4.7"
xmltodict = "^0.13.0"
Markdown = "^3.3.7"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
drf-yasg = "^1.20.0"
flex = "^6.14.1"
django-extensions = "^3.1.5"
ipython = "^8.4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
